9/17/2019 Images Update
Created multiple favicons depending on what device the site is being viewed from. Obviously once this is all said and done, I'm not going to use the name "FantArby's", but until I actually release this and come up with a better (non-copyrighted) name, this is how it's going to be.