9/17/2019 Includes Update
Added the filltable.inc.php file, which is an extension of the developer php file in the main folder. I've been attempting to make this file upload information from a JSON file from The Blue Alliance APIs to the database on my host site. It does not currently work, but I'm making decent progress. Some of the other files in this folder were also updated due to updates to the database.